refactor: move related components to `Explorer.tsx`
